Node selection fixes
