+++ package.json
@@ -54,7 +54,7 @@
         "vite-tsconfig-paths": "^4.2.0"
     },
     "dependencies": {
-        "@powerhousedao/design-system": "^1.0.0-alpha.16",
+        "@powerhousedao/design-system": "^1.0.0-alpha.17",
         "@typescript-eslint/eslint-plugin": "^5.59.5",
         "document-model": "^1.0.8",
         "electron-deeplink": "^1.0.10",
